+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/BaseIntegrationTest.java
@@ -23,6 +23,7 @@ import org.junit.BeforeClass;
 
 import io.realm.Realm;
 import io.realm.SyncManager;
+import io.realm.log.LogLevel;
 import io.realm.log.RealmLog;
 import io.realm.objectserver.utils.HttpUtils;
 
@@ -33,6 +34,7 @@ class BaseIntegrationTest {
         SyncManager.Debug.skipOnlineChecking = true;
         try {
             Realm.init(InstrumentationRegistry.getContext());
+            RealmLog.setLevel(LogLevel.DEBUG);
             HttpUtils.startSyncServer();
         } catch (Exception e) {
             // Throwing an exception from this method will crash JUnit. Instead just log it.
@@ -45,6 +47,7 @@ class BaseIntegrationTest {
     public static void tearDown () throws Exception {
         try {
             HttpUtils.stopSyncServer();
+            RealmLog.setLevel(LogLevel.WARN);
         } catch (Exception e) {
             RealmLog.error("Failed to stop Sync Server", e);
         }